+++ clinicadl/commandline/pipelines/generate/shepplogan/cli.py
@@ -9,12 +9,12 @@ from joblib import Parallel, delayed
 from clinicadl.commandline import arguments
 from clinicadl.commandline.modules_options import data, dataloader
 from clinicadl.commandline.pipelines.generate.shepplogan import options as shepplogan
+from clinicadl.data.datatype.file_type import FileType
 from clinicadl.generate.generate_config import GenerateSheppLoganConfig
 from clinicadl.generate.generate_utils import (
     generate_shepplogan_phantom,
     write_missing_mods,
 )
-from clinicadl.utils.iotools.clinica_utils import FileType
 from clinicadl.utils.iotools.iotools import check_and_clean, commandline_to_json
 from clinicadl.utils.iotools.utils import write_preprocessing
