--- clinicadl/data/datatype/__init__.py
@@ -1,1 +1,0 @@
-from .file_type import FileType
