+++ clinicadl/data/datatype/modalities/custom.py
@@ -1,6 +1,8 @@
 from pydantic import computed_field, field_validator
 
-from .base import ImageModality, Modality
+from clinicadl.data.datatype.utils import ImageModality
+
+from .base import Modality
 
 
 class Custom(Modality):
