+++ clinicadl/data/datatype/modalities/pet.py
@@ -1,8 +1,10 @@
 from enum import Enum
 
-from pydantic import computed_field, field_validator
+from pydantic import computed_field
 
-from .base import ImageModality, Modality
+from clinicadl.data.datatype.utils import ImageModality
+
+from .base import Modality
 
 
 class Tracer(str, Enum):
