+++ clinicadl/data/datatype/preprocessing/custom.py
@@ -2,8 +2,9 @@ from pydantic import computed_field
 
 from clinicadl.data.datatype.file_type import FileType
 from clinicadl.data.datatype.modalities import Custom as CustomModality
+from clinicadl.data.datatype.utils import PreprocessingMethod
 
-from .base import Preprocessing, PreprocessingMethod
+from .base import Preprocessing
 
 
 class Custom(Preprocessing, CustomModality):
