+++ clinicadl/data/datatype/preprocessing/pet.py
@@ -22,9 +22,15 @@ class PETLinear(_PreprocessingWithCrop, PET):
 
         des_crop = "" if self.use_uncropped_image else "_desc-Crop"
 
+        description = f"{self.modality.value} Image registered in MNI152NLin2009cSym space using pet-linear pipeline"
+
+        if not self.use_uncropped_image:
+            description += (
+                " and cropped (matrix size 169×208×179, 1 mm isotropic voxels)"
+            )
         return FileType(
             pattern=f"pet_linear/*_trc-{self.tracer}_space-MNI152NLin2009cSym{des_crop}_res-1x1x1_suvr-{self.suvr_reference_region}_pet.nii.gz",
-            description="",
+            description=description,
             needed_pipeline=PreprocessingMethod.PET_LINEAR,
         )
