+++ clinicadl/data/datatype/utils.py
@@ -1,0 +1,21 @@
+from enum import Enum
+
+
+class ImageModality(str, Enum):
+    """Possible modality for images in clinicaDL."""
+
+    T1W = "T1w"
+    DWI = "dwi"
+    PET = "pet"
+    FLAIR = "flair"
+    CUSTOM = "custom"
+
+
+class PreprocessingMethod(str, Enum):
+    """Possible preprocessing methods available in Clinica."""
+
+    T1_LINEAR = "t1-linear"
+    PET_LINEAR = "pet-linear"
+    FLAIR_LINEAR = "flair-linear"
+    CUSTOM = "custom"
+    DWI_DTI = "dwi-dti"
