--- tests/unittests/dataset/datatype/modalities/__init__.py
@@ -1,6 +1,0 @@
-from .base import ImageModality
-from .custom import Custom
-from .dwi import DWI
-from .flair import Flair
-from .pet import PET
-from .t1w import T1w
