--- tests/unittests/dataset/datatype/preprocessing/__init__.py
@@ -1,6 +1,0 @@
-from .base import Preprocessing, PreprocessingMethod
-from .custom import Custom
-from .dti import DWIDTI
-from .flair import FlairLinear
-from .pet import PETLinear
-from .t1 import T1Linear
