--- tests/unittests/dataset/datatype/raw/__init__.py
@@ -1,6 +1,0 @@
-from .base import RawData
-from .custom import RawCustom
-from .dwi import RawDWI
-from .flair import RawFlair
-from .pet import RawPET
-from .t1w import RawT1w
